Fix credentials file created with 'undefined'.
There were two problems here.

First we were calling `writeFile` with no
contents parameter which resulted in putting the string "undefined" into
the contents of the file.

Second we were calling `writeFile` which is asynchronous and not waiting
on the promise and proceeding as if it was synchronous.

Lastly this creates this hidden folder with restricted permissions since it
does contain credentials.