Remove implicit dependency between IAM policy and log groups
Until now, the IAM policy had a dependency on log groups completing
before it could finish because of the use of `Fn::GetAtt` with a log
group, meaning the policy couldn't provision until all log groups were
ready.

This patch changes the log group reference in the CFN template from:

```
{ "Fn::GetAtt": ["MyFuncLogGroup", "Arn"] }
```

to

```
{"Fn::Sub": "arn:aws:logs:${AWS::Region}:${AWS::AccountId}:log-group:/aws/lambda/my-service-dev-myfunction:*"}
```

Meaning the policy can be resolved 100% with pseudoparams for log group ARNs.
This speeds up the CFN initial deploy significantly, and somewhat
improves deployments that add or rename functions.

Having this dependency removed is also the first step in allowing log
groups to be moved to a nested stack.