Improves usage and documentation of SNS event sources.
When using cloudformation intrinsic functions, it is easy to assume that sns:topicName is a property that should resolve to the actual value in AWS. In reality it is only used to name the underlying cloudformation resources generated by serverless, and thus needs to be a string.
This improves the error message and code path returned when sns:topicName is not a string, as well as does some validation around usage of Fn::ImportValue. Also adds some tests and documentation demonstrating that intrinsic functions work with SNS event sources.